project 2/requirements.txt
Flask
google-generativeai
PyPDF2
python-docx
werkzeug
gunicorn


